requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
